+++ tests/Jroc.Test262.Tests/built-ins/TypedArray/prototype/forEach/BigInt/ExecutionTests.cs
@@ -1,0 +1,20 @@
+using Jroc.Test262.Tests.built_ins;
+
+namespace Jroc.Test262.Tests.built_ins.TypedArray.prototype.forEach.BigInt;
+
+public class ExecutionTests : DiskExecutionTestsBase
+{
+    public ExecutionTests() : base("built_ins.TypedArray.prototype.forEach.BigInt") { }
+
+    [Fact(DisplayName = "callbackfn-detachbuffer")]
+    public Task callbackfn_detachbuffer()
+        => ExecutionTestFromFile("callbackfn-detachbuffer");
+
+    [Fact(DisplayName = "detached-buffer")]
+    public Task detached_buffer()
+        => ExecutionTestFromFile("detached-buffer");
+
+    [Fact(DisplayName = "return-abrupt-from-this-out-of-bounds")]
+    public Task return_abrupt_from_this_out_of_bounds()
+        => ExecutionTestFromFile("return-abrupt-from-this-out-of-bounds");
+}